refactor: improve code redability and remove unused components and pages
