--- frontend/src/data/mockedGyms.js
@@ -1,145 +1,0 @@
-const gyms = [
-  {
-    name: 'Gym A',
-    workouts: [
-      {
-        "id": "a9c39103-a0a6-467f-b833-8da3a64982a8",
-        "name": "Two Exercises",
-        "status": "active",
-        "updated_at": "2025-05-08T10:54:26.147164Z",
-        "gym_id": "f1dd41bd-4012-43df-9b93-866120c0dfae",
-        "exerciseList": [
-          {
-            "id": "1bd2ebe4-bbd9-4b71-b2f3-d07611981869",
-            "name": "ccc",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": "Do it every Wednesday",
-            "status": "active",
-            "workout_id": "a9c39103-a0a6-467f-b833-8da3a64982a8",
-            "updated_at": "2025-05-08T10:54:26.187381Z"
-          },
-          {
-            "id": "64628c61-e939-4ef2-99dc-e5ffcd8fd84b",
-            "name": "ddd",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": "Never skip",
-            "status": "active",
-            "workout_id": "a9c39103-a0a6-467f-b833-8da3a64982a8",
-            "updated_at": "2025-05-08T10:54:26.191381Z"
-          }
-        ]
-      },
-      {
-        "id": "adbc0065-a02f-4da4-bb31-652e55ffa0da",
-        "name": "Workout With No exercises",
-        "status": "active",
-        "updated_at": "2025-05-08T10:58:20.469363Z",
-        "gym_id": "f1dd41bd-4012-43df-9b93-866120c0dfae",
-        "exerciseList": []
-      },
-      {
-        "id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-        "name": "Workout eight exercises",
-        "status": "active",
-        "updated_at": "2025-05-08T11:01:38.925072Z",
-        "gym_id": "f1dd41bd-4012-43df-9b93-866120c0dfae",
-        "exerciseList": [
-          {
-            "id": "79838115-efc4-4978-a8b9-b86a03978812",
-            "name": "aaa",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T10:53:00.462711Z"
-          },
-          {
-            "id": "5142ea78-b513-4250-9a4e-ae9607f14684",
-            "name": "bbb",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T10:53:00.463872Z"
-          },
-          {
-            "id": "662be9a6-e304-4314-b9ae-b8bafe55d03a",
-            "name": "ddd",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T11:01:20.203153Z"
-          },
-          {
-            "id": "4ad771ce-6345-4db7-a656-69fe09264b48",
-            "name": "ccc",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T11:01:20.206135Z"
-          },
-          {
-            "id": "fcb4b9a7-e8b2-4543-8b70-6d6e9acfe280",
-            "name": "ddd",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T11:01:34.926425Z"
-          },
-          {
-            "id": "2b8c4b44-27d3-4e71-8119-08b182e7dd1f",
-            "name": "ccc",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T11:01:34.929414Z"
-          },
-          {
-            "id": "4d56d3cb-a8dc-409c-96b9-45231f1428f7",
-            "name": "ddd",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T11:01:38.920071Z"
-          },
-          {
-            "id": "b8c81f01-93bf-497c-89b7-de6ed360c08f",
-            "name": "ccc",
-            "weight": 1.0,
-            "numReps": 1,
-            "numSets": 2,
-            "note": null,
-            "status": "active",
-            "workout_id": "2f224c4c-8c7a-44e0-918c-bddce9d6c4d6",
-            "updated_at": "2025-05-08T11:01:38.923071Z"
-          }
-        ]
-      }
-    ]
-  }
-];
-
-export default gyms;
