+++ frontend/src/pages/HomePage.jsx
@@ -1,13 +1,12 @@
-import React, { useState } from 'react';
-import { Box, Typography, Dialog, useMediaQuery, useTheme, Alert, Snackbar } from '@mui/material';
+import React from 'react';
+import { Box, Typography } from '@mui/material';
 import GreetingTitle from '../components/GreetingTitle';
 import GymHeader from '../components/GymHeader';
 import WorkoutsList from '../components/WorkoutsList';
-import WorkoutCard from '../components/WorkoutCard';
-import WorkoutCardActions from '../components/WorkoutCardActions';
+import WorkoutDialog from '../components/WorkoutDialog';
 import { useGymManagement } from '../hooks/useGymManagement';
 import { useWorkoutsByGymId } from '../hooks/useWorkoutsByGymId';
-import { deleteWorkout, updateWorkout, createWorkout } from '../api/WorkoutApi';
+import { useWorkoutManagement } from '../hooks/useWorkoutManagement';
 import mockedUsers from '../data/mockedUsers';
 
 const user = mockedUsers[0];
@@ -31,98 +30,23 @@ export default function HomePage() {
 
   const { workouts, loading: workoutsLoading, error: workoutsError, refresh: refreshWorkouts } = useWorkoutsByGymId(selectedGym?.id);
 
-  const [openWorkout, setOpenWorkout] = useState(null);
-  const [isEditingWorkout, setIsEditingWorkout] = useState(false);
-  const theme = useTheme();
-  const fullScreen = useMediaQuery(theme.breakpoints.down('sm'));
-  const [deleteError, setDeleteError] = useState(null);
-  const [isDeleting, setIsDeleting] = useState(false);
-  const [editedWorkout, setEditedWorkout] = useState(null);
-  const [isUpdating, setIsUpdating] = useState(false);
-  const [isCreatingWorkout, setIsCreatingWorkout] = useState(false);
-
-  const handleCreateWorkoutClick = () => {
-    setIsCreatingWorkout(true);
-    setOpenWorkout(null);
-    setEditedWorkout({
-      name: 'Name your new workout',
-      exerciseList: [{ name: '', weight: 0, numReps: 0, numSets: 0, note: '' }]
-    });
-    setIsEditingWorkout(true);
-  };
-
-  const handleOpenWorkout = (workout) => {
-    setOpenWorkout(workout);
-    setEditedWorkout(workout);
-    setIsEditingWorkout(true);
-    setIsCreatingWorkout(false);
-  };
-
-  const handleCloseWorkout = () => {
-    setOpenWorkout(null);
-    setEditedWorkout(null);
-    setIsEditingWorkout(false);
-    setIsCreatingWorkout(false);
-  };
-
-  const handleEditWorkout = () => {
-    setIsEditingWorkout(true);
-    setEditedWorkout(openWorkout);
-  };
-
-  const handleWorkoutChange = (updatedWorkout) => {
-    setEditedWorkout(updatedWorkout);
-  };
-
-  const handleSaveWorkout = async () => {
-    if (!editedWorkout) return;
-    try {
-      setIsUpdating(true);
-      if (isCreatingWorkout) {
-        await createWorkout(selectedGym.id, {
-          name: editedWorkout.name,
-          exerciseList: editedWorkout.exerciseList
-        });
-      } else {
-        await updateWorkout(editedWorkout.id, {
-          name: editedWorkout.name,
-          exerciseList: editedWorkout.exerciseList
-        });
-      }
-      setIsEditingWorkout(false);
-      setIsCreatingWorkout(false);
-      handleCloseWorkout();
-      refreshWorkouts();
-    } catch (error) {
-      const errorMsg = error.response?.data?.errors?.[0] || error.response?.data?.message || 'Failed to save workout';
-      alert(errorMsg);
-    } finally {
-      setIsUpdating(false);
-    }
-  };
-
-  const handleCancelWorkout = () => {
-    setEditedWorkout(openWorkout);
-    setIsEditingWorkout(false);
-  };
-
-  const handleDeleteWorkout = async () => {
-    if (!openWorkout) return;
-    
-    const confirmed = window.confirm('Are you sure you want to delete this workout?');
-    if (!confirmed) return;
-
-    try {
-      setIsDeleting(true);
-      await deleteWorkout(openWorkout.id);
-      handleCloseWorkout();
-      refreshWorkouts();
-    } catch (error) {
-      setDeleteError(error.response?.data?.message || 'Failed to delete workout');
-    } finally {
-      setIsDeleting(false);
-    }
-  };
+  const {
+    openWorkout,
+    isEditingWorkout,
+    isCreatingWorkout,
+    editedWorkout,
+    isUpdating,
+    isDeleting,
+    deleteError,
+    handleCreateWorkoutClick,
+    handleOpenWorkout,
+    handleCloseWorkout,
+    handleEditWorkout,
+    handleWorkoutChange,
+    handleSaveWorkout,
+    handleCancelWorkout,
+    handleDeleteWorkout
+  } = useWorkoutManagement(selectedGym?.id, refreshWorkouts);
 
   return (
     <Box>
@@ -175,50 +99,19 @@ export default function HomePage() {
         )}
       </Box>
 
-      <Dialog
+      <WorkoutDialog
         open={!!openWorkout || isCreatingWorkout}
         onClose={handleCloseWorkout}
-        fullScreen={fullScreen}
-        maxWidth="md"
-        fullWidth
-        aria-labelledby="workout-dialog-title"
-        PaperProps={{
-          sx: {
-            background: 'none',
-            boxShadow: 'none',
-            overflow: 'visible',
-            m: { xs: 1, sm: 3 },
-            maxHeight: 'calc(100vh - 32px)',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-          }
-        }}
-      >
-        <Box sx={{ width: '100%', maxWidth: { xs: '100%', md: 800 }, display: 'flex', flexDirection: 'column', alignItems: 'center', height: '100%' }}>
-          <Box sx={{ width: '100%', px: 2, pt: 2, pb: 1, position: 'sticky', top: 0, zIndex: 2, background: 'background.paper', borderTopLeftRadius: 16, borderTopRightRadius: 16 }}>
-            <WorkoutCardActions
-              isEditing={isEditingWorkout}
-              onEdit={handleEditWorkout}
-              onSave={handleSaveWorkout}
-              onCancel={handleCancelWorkout}
-              onDelete={handleDeleteWorkout}
-              onClose={handleCloseWorkout}
-              isDeleting={isDeleting}
-              isUpdating={isUpdating}
-            />
-          </Box>
-          <Box sx={{ width: '100%', flex: 1, overflow: 'auto', display: 'flex', justifyContent: 'center', alignItems: 'flex-start', maxHeight: 'calc(100vh - 120px)' }}>
-            {editedWorkout && (
-              <WorkoutCard 
-                workout={editedWorkout} 
-                isEditing={isEditingWorkout}
-                onWorkoutChange={handleWorkoutChange}
-              />
-            )}
-          </Box>
-        </Box>
-      </Dialog>
+        editedWorkout={editedWorkout}
+        isEditingWorkout={isEditingWorkout}
+        isDeleting={isDeleting}
+        isUpdating={isUpdating}
+        onEdit={handleEditWorkout}
+        onSave={handleSaveWorkout}
+        onCancel={handleCancelWorkout}
+        onDelete={handleDeleteWorkout}
+        onWorkoutChange={handleWorkoutChange}
+      />
     </Box>
   );
 }
